Complete and fully responsive hero section
