+++ src/app/home-page/components/hero-section/hero-section.component.css
@@ -1,0 +1,20 @@
+.video {
+    background-repeat: no-repeat;
+    background-size: cover;
+    background-position: center;
+    width: 100%;
+    height: 70vh;
+    object-fit: cover;
+}
+
+.box-area {
+    display: grid;
+    grid-template-columns: repeat(auto-fit, minmax(160px, 1fr));
+    grid-gap: 22px;
+}
+
+@media (max-width: 1107px) {
+  .box-area > *:last-child:nth-child(odd) {
+    grid-column: 1 / -1;
+  }
+}
