+++ src/app/home-page/components/hero-section/hero-section.component.ts
@@ -1,0 +1,11 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-hero-section',
+  imports: [],
+  templateUrl: './hero-section.component.html',
+  styleUrl: './hero-section.component.css'
+})
+export class HeroSectionComponent {
+
+}
