+++ src/app/home-page/home-page.component.html
@@ -1,3 +1,5 @@
 <div class="w-full h-auto flex flex-col">
     <app-navbar></app-navbar>
+
+    <app-hero-section></app-hero-section>
 </div>
