+++ src/app/home-page/home-page.component.ts
@@ -1,10 +1,12 @@
 import { Component } from '@angular/core';
-import { NavbarComponent } from './navbar/navbar.component';
+import { NavbarComponent } from './components/navbar/navbar.component';
+import { HeroSectionComponent } from './components/hero-section/hero-section.component';
 
 @Component({
   selector: 'app-home-page',
   imports: [
-    NavbarComponent
+    NavbarComponent,
+    HeroSectionComponent
   ],
   templateUrl: './home-page.component.html',
   styleUrl: './home-page.component.css'
